chore: Move src/config/anonymousCommunityChatroomSocketHandlers to src/socketHandlers/anonCommunityChat
